chain: move subscriber to internal node package
